Changed byte to int to reduce redundancy

+++ src/game/Curved.java
@@ -2,11 +2,11 @@ package game;
 
 public class Curved extends Pipe {
 
-    public Curved(byte x, byte y, boolean upEdge, boolean downEdge, boolean leftEdge, boolean rightEdge) {
+    public Curved(int x, int y, boolean upEdge, boolean downEdge, boolean leftEdge, boolean rightEdge) {
         super(x, y, true, upEdge, downEdge, leftEdge, rightEdge);
     }
 
-    public Curved(byte x, byte y, byte position1, byte position2) {
+    public Curved(int x, int y, int position1, int position2) {
 
         super(x, y, false, false, false, false, false);
         setPosition(position1, position2);
